Tag服务添加用户权限检查

+++ src/main/java/top/lizec/smartreview/service/TagService.java
@@ -26,6 +26,7 @@ public class TagService {
 
     @Transactional
     public void remove(Integer tagId, Integer creator) {
+        checkUserPermission(creator, tagId);
         Tag tag = new Tag();
         tag.setId(tagId);
         tagDao.delete(tag);
